Remove global includes from Common.Legacy
Replace macro atol with std.

+++ src/shared/Database/Field.hpp
@@ -15,15 +15,15 @@ public:
     void SetValue(char* value) { mValue = value; }
 
     const char* GetString() { return mValue; }
-    float GetFloat() { return mValue ? static_cast<float>(atof(mValue)) : 0; }
-    bool GetBool() { return mValue ? atoi(mValue) > 0 : false; }
+    float GetFloat() { return mValue ? static_cast<float>(std::stod(mValue)) : 0; }
+    bool GetBool() { return mValue ? std::stoi(mValue) > 0 : false; }
 
-    uint8_t GetUInt8() { return mValue ? static_cast<uint8_t>(atol(mValue)) : 0U; }
-    int8_t GetInt8() { return mValue ? static_cast<int8_t>(atol(mValue)) : 0; }
-    uint16_t GetUInt16() { return mValue ? static_cast<uint16_t>(atol(mValue)) : 0U; }
-    int16_t GetInt16() { return mValue ? static_cast<int16_t>(atol(mValue)) : 0; }
-    uint32_t GetUInt32() { return mValue ? static_cast<uint32_t>(atol(mValue)) : 0U; }
-    int32_t GetInt32() { return mValue ? static_cast<int32_t>(atol(mValue)) : 0; }
+    uint8_t GetUInt8() { return mValue ? static_cast<uint8_t>(std::stoul(mValue)) : 0U; }
+    int8_t GetInt8() { return mValue ? static_cast<int8_t>(std::stoi(mValue)) : 0; }
+    uint16_t GetUInt16() { return mValue ? static_cast<uint16_t>(std::stoul(mValue)) : 0U; }
+    int16_t GetInt16() { return mValue ? static_cast<int16_t>(std::stoi(mValue)) : 0; }
+    uint32_t GetUInt32() { return mValue ? static_cast<uint32_t>(std::stoul(mValue)) : 0U; }
+    int32_t GetInt32() { return mValue ? static_cast<int32_t>(std::stoi(mValue)) : 0; }
 
     uint64_t GetUInt64()
     {
